+++ lib/mosquitto.c
@@ -327,7 +327,7 @@ void mosquitto_destroy(struct mosquitto *mosq)
 int mosquitto_socket(struct mosquitto *mosq)
 {
 	if(!mosq) return INVALID_SOCKET;
-	return mosq->sock;
+	return (int)mosq->sock;
 }
